+++ R/add-tbl.R
@@ -38,7 +38,7 @@ dm_add_tbl <- function(dm, ..., repair = "unique", quiet = FALSE) {
 }
 
 repair_names_vec <- function(names, repair, quiet) {
-  tryCatch(
+  withCallingHandlers(
     vctrs::vec_as_names(names, repair = repair, quiet = quiet),
     vctrs_error_names_must_be_unique = function(e) {
       abort_need_unique_names(names[duplicated(names)])
@@ -82,7 +82,7 @@ dm_add_tbl_impl <- function(dm, tbls, table_name, filters = vctrs::list_of(new_f
 #' `tidyselect` is supported, see [`dplyr::select()`] for details on the semantics.
 #'
 #' @export
-#' @examples
+#' @examplesIf rlang::is_installed("nycflights13")
 #' dm_nycflights13() %>%
 #'   dm_rm_tbl(airports)
 dm_rm_tbl <- function(dm, ...) {