+++ R/db-interface.R
@@ -14,39 +14,77 @@
 #'
 #' @param dest An object of class `"src"` or `"DBIConnection"`.
 #' @param dm A `dm` object.
-#' @param table_names A named character vector, containing the names that you want the tables in the `dm` to have
-#'   after copying them to the database.
-#'   The table names within the `dm` will remain unchanged.
-#'   The name of each element of the vector needs to be one of the table names of the `dm`.
-#'   Those tables of the `dm` that are not addressed will be called by their original name on the database.
 #' @param overwrite,types,indexes,unique_indexes Must remain `NULL`.
 #' @param set_key_constraints Boolean variable, if `TRUE` will mirror `dm` key constraints on a database.
-#' @param unique_table_names Boolean, if `FALSE` (default), the original table names will be used, if `TRUE`,
-#'   unique table names will be created based on the original table names.
+#' @param unique_table_names Deprecated.
 #' @param temporary Boolean variable, if `TRUE`, only temporary tables will be created.
 #'   These tables will vanish when disconnecting from the database.
-#' @param ... Possible further arguments passed to [dplyr::copy_to()], which is used on each table.
+#' @param table_names Desired names for the tables on `dest`; the names within the `dm` remain unchanged.
+#'   Can be `NULL`, a named character vector, a function or a one-sided formula.
+#'
+#'   If left `NULL` (default), the names will be determined automatically depending on the `temporary` argument:
+#'
+#'   1. `temporary = TRUE` (default): unique table names based on the names of the tables in the `dm` are created.
+#'   1. `temporary = FALSE`: the table names in the `dm` are used as names for the tables on `dest`.
+#'
+#'   If a function or one-sided formula, `table_names` is converted to a function
+#'   using [rlang::as_function()].
+#'   This function is called with the unquoted table names of the `dm` object
+#'   as the only argument.
+#'   The output of this function is processed by [DBI::dbQuoteIdentifier()],
+#'   that result should be a vector of identifiers of the same length
+#'   as the original table names.
+#'
+#'   Use a variant of
+#'   `table_names = ~ DBI::SQL(dbplyr::in_schema("schema_name", .x))`
+#'   to specify the same schema for all tables.
+#'   Use `table_names = identity` with `temporary = TRUE`
+#'   to avoid giving temporary tables unique names.
+#'
+#'   If a named character vector,
+#'   the names of this vector need to correspond to the table names in the `dm`,
+#'   and its values are the desired names on `dest`.
+#'   The value is processed by [DBI::dbQuoteIdentifier()],
+#'   that result should be a vector of identifiers of the same length
+#'   as the original table names.
+#'
+#'   Use qualified names corresponding to your database's syntax
+#'   to specify e.g. database and schema for your tables.
+#' @param ... Passed on to [dplyr::copy_to()], which is used on each table.
 #'
 #' @family DB interaction functions
 #'
-#' @return A `dm` object on the given `src`.
+#' @return A `dm` object on the given `src` with the same table names
+#'   as the input `dm`.
 #'
-#' @examples
-#' src_sqlite <- dplyr::src_sqlite(":memory:", create = TRUE)
-#' iris_dm <- copy_dm_to(
-#'   src_sqlite,
-#'   as_dm(list(iris = iris)),
+#' @examplesIf rlang::is_installed("RSQLite") && rlang::is_installed("nycflights13") && rlang::is_installed("dbplyr")
+#' con <- DBI::dbConnect(RSQLite::SQLite())
+#'
+#' # Copy to temporary tables, unique table names by default:
+#' temp_dm <- copy_dm_to(
+#'   con,
+#'   dm_nycflights13(),
 #'   set_key_constraints = FALSE
 #' )
+#'
+#' # Persist, explicitly specify table names:
+#' persistent_dm <- copy_dm_to(
+#'   con,
+#'   dm_nycflights13(),
+#'   temporary = FALSE,
+#'   table_names = ~ paste0("flights_", .x)
+#' )
+#' dbplyr::remote_name(persistent_dm$planes)
+#'
+#' DBI::dbDisconnect(con)
 #' @export
 copy_dm_to <- function(dest, dm, ...,
                        types = NULL, overwrite = NULL,
                        indexes = NULL, unique_indexes = NULL,
-                       set_key_constraints = TRUE, unique_table_names = FALSE,
+                       set_key_constraints = TRUE, unique_table_names = NULL,
                        table_names = NULL,
                        temporary = TRUE) {
   # for the time being, we will be focusing on MSSQL
-  # we expect the src (dest) to already point to the correct schema
   # we want to
   #   1. change `dm_get_src(dm)` to `dest`
   #   2. copy the tables to `dest`
@@ -68,16 +106,61 @@ copy_dm_to <- function(dest, dm, ...,
     abort_no_unique_indexes()
   }
 
-  if (!is.null(table_names)) {
-    if (unique_table_names) {
-      abort_unique_table_names_or_table_names()
+  if (!is.null(unique_table_names)) {
+    deprecate_soft(
+      "0.1.4", "dm::copy_dm_to(unique_table_names = )",
+      details = "Use `table_names = identity` to use unchanged names for temporary tables."
+    )
+
+    if (is.null(table_names) && temporary && !unique_table_names) {
+      table_names <- identity
+    }
+  }
+
+  dest <- src_from_src_or_con(dest)
+  src_names <- src_tbls(dm)
+
+  if (is_db(dest)) {
+    dest_con <- con_from_src_or_con(dest)
+
+    # in case `table_names` was chosen by the user, check if the input makes sense:
+    # 1. is there one name per dm-table?
+    # 2. are there any duplicated table names?
+    # 3. is it a named character or ident_q vector with the correct names?
+    if (is.null(table_names)) {
+      table_names_out <- repair_table_names_for_db(src_names, temporary, dest_con)
+
+      # https://github.com/tidyverse/dbplyr/issues/487
+      if (is_mssql(dest)) {
+        temporary <- FALSE
+      }
+    } else {
+      if (is_function(table_names) || is_bare_formula(table_names)) {
+        table_name_fun <- as_function(table_names)
+        table_names_out <- set_names(table_name_fun(src_names), src_names)
+      } else {
+        table_names_out <- table_names
+      }
+      check_naming(names(table_names_out), src_names)
+
+      if (anyDuplicated(table_names_out)) {
+        problem <- table_names_out[duplicated(table_names_out)][[1]]
+        abort_copy_dm_to_table_names_duplicated(problem)
+      }
+
+      table_names_out <- unclass(DBI::dbQuoteIdentifier(dest_con, table_names_out[src_names]))
+      # names(table_names_out) <- src_names
     }
 
-    not_found <- setdiff(names2(table_names), src_tbls(dm))
-    if (has_length(not_found)) {
-      if (any(not_found == "")) abort_need_named_vec(src_tbls(dm))
-      abort_table_not_in_dm(unique(not_found), src_tbls(dm))
-    }
+    # create `ident`-class objects from the table names
+    table_names_out <- map(table_names_out, dbplyr::ident_q)
+  } else {
+    # FIXME: Other data sources than local and database possible
+    deprecate_soft(
+      "0.1.6", "dm::copy_dm_to(dest = 'must refer to a remote data source')",
+      "dm::collect.dm()"
+    )
+    table_names_out <- set_names(src_names)
   }
 
   check_not_zoomed(dm)
@@ -85,10 +168,14 @@ copy_dm_to <- function(dest, dm, ...,
   # FIXME: if same_src(), can use compute() but need to set NOT NULL
   # constraints
 
-  dest <- src_from_src_or_con(dest)
   dm <- collect(dm)
 
-  copy_data <- build_copy_data(dm, dest, table_names, unique_table_names)
+  # Shortcut necessary to avoid copying into .GlobalEnv
+  if (!is_db(dest)) {
+    return(dm)
+  }
+
+  copy_data <- build_copy_data(dm, dest, table_names_out)
 
   new_tables <- copy_list_of_tables_to(
     dest,
@@ -121,18 +208,19 @@ copy_dm_to <- function(dest, dm, ...,
 #'
 #' @return Returns the `dm`, invisibly. Side effect: installing key constraints on DB.
 #'
-#' @examples
-#' src_sqlite <- dplyr::src_sqlite(":memory:", create = TRUE)
-#' iris_dm <- copy_dm_to(
-#'   src_sqlite,
-#'   as_dm(list(iris = iris)),
+#' @examplesIf rlang::is_installed("RSQLite") && rlang::is_installed("nycflights13")
+#' # Setting key constraints not yet supported on SQLite,
+#' # try this with SQL Server or Postgres instead:
+#' sqlite <- DBI::dbConnect(RSQLite::SQLite())
+#'
+#' flights_dm <- copy_dm_to(
+#'   sqlite,
+#'   dm_nycflights13(),
 #'   set_key_constraints = FALSE
 #' )
 #'
-#' # there are no key constraints in `as_dm(list(iris = iris))`
-#' # but if there were, and if we had already implemented setting key
-#' # constraints for SQLite, the following command would do something:
-#' dm_set_key_constraints(iris_dm)
+#' dm_set_key_constraints(flights_dm)
+#' DBI::dbDisconnect(sqlite)
 #' @noRd
 dm_set_key_constraints <- function(dm) {
   if (!is_src_db(dm) && !is_this_a_test()) abort_key_constraints_need_db()
@@ -153,4 +241,42 @@ dm_set_key_constraints <- function(dm) {
   walk(queries, ~ dbExecute(con, .))
 
   invisible(dm)
+}
+
+get_db_table_names <- function(dm) {
+  if (!is_src_db(dm)) {
+    return(tibble(table_name = src_tbls(dm), remote_name = src_tbls(dm)))
+  }
+  tibble(
+    table_name = src_tbls(dm),
+    remote_name = map_chr(dm_get_tables_impl(dm), dbplyr::remote_name)
+  )
+}
+
+check_naming <- function(table_names, dm_table_names) {
+  if (!identical(sort(table_names), sort(dm_table_names))) {
+    abort_copy_dm_to_table_names()
+  }
+}
+
+
+# Errors ------------------------------------------------------------------
+
+abort_copy_dm_to_table_names <- function() {
+  abort(error_txt_copy_dm_to_table_names(), .subclass = dm_error_full("copy_dm_to_table_names"))
+}
+
+error_txt_copy_dm_to_table_names <- function() {
+  "`table_names` must have names that are the same as the table names in `dm`."
+}
+
+abort_copy_dm_to_table_names_duplicated <- function(problem) {
+  abort(error_txt_copy_dm_to_table_names_duplicated(problem), .subclass = dm_error_full("copy_dm_to_table_names_duplicated"))
+}
+
+error_txt_copy_dm_to_table_names_duplicated <- function(problem) {
+  c(
+    "`table_names` must be unique.",
+    i = paste0("Duplicate: ", tick(problem))
+  )
 }