+++ R/dm.R
@@ -33,9 +33,10 @@
 #'
 #' @export
 #' @examples
-#' dm(iris, mtcars)
-#' new_dm(list(iris = iris, mtcars = mtcars))
-#' as_dm(list(iris = iris, mtcars = mtcars))
+#' dm(trees, mtcars)
+#' new_dm(list(trees = trees, mtcars = mtcars))
+#' as_dm(list(trees = trees, mtcars = mtcars))
+#' @examplesIf rlang::is_installed("nycflights13") && rlang::is_installed("dbplyr")
 #'
 #' dm_nycflights13() %>% tbl("airports")
 #' dm_nycflights13() %>% src_tbls()
@@ -63,7 +64,7 @@ dm <- function(..., .name_repair = c("check_unique", "unique", "universal", "min
   if (has_length(quos)) {
     src_index <- c(which(names(quos) == "src"), 1)[[1]]
     if (is.src(tbls[[src_index]])) {
-      lifecycle::deprecate_soft("0.0.4.9001", "dm::dm(src = )", "dm_from_src()")
+      deprecate_soft("0.0.4.9001", "dm::dm(src = )", "dm_from_src()")
       return(invoke(dm_from_src, tbls))
     }
   }
@@ -243,7 +244,8 @@ debug_validate_dm <- function(dm) {
 #'
 #' @rdname dm
 #'
-#' @return For `dm_get_src()`: the \pkg{dplyr} source for a `dm` object.
+#' @return For `dm_get_src()`: the \pkg{dplyr} source for a `dm` object,
+#'   or `NULL` if the `dm` objcet contains data frames.
 #'
 #' @export
 dm_get_src <- function(x) {
@@ -440,9 +442,9 @@ as_dm.default <- function(x) {
 
 tbl_src <- function(x) {
   if (is_empty(x) || is.data.frame(x)) {
-    default_local_src()
+    NULL
   } else if (inherits(x, "tbl_sql")) {
-    x$src
+    dbplyr::remote_src(x)
   } else {
     abort_what_a_weird_object(class(x)[[1]])
   }
@@ -454,30 +456,39 @@ as_dm.src <- function(x) {
 }
 
 #' @export
-format.dm <- function(x, ...) {
-  def <- dm_get_def(x)
-  glue("dm: {def_get_n_tables(def)} tables, {def_get_n_columns(def)} columns, {def_get_n_pks(def)} primary keys, {def_get_n_fks(def)} foreign keys")
+print.dm <- function(x, ...) { # for both dm and zoomed_dm
+  show_dm(x)
+  invisible(x)
 }
 
 #' @export
-print.dm <- function(x, ...) {
-  cat_rule("Table source", col = "green")
-  def <- dm_get_def(x)
-  src <- dm_get_src(x)
-  db_info <- NULL
+print.zoomed_dm <- function(x, ..., n = NULL, width = NULL, n_extra = NULL) {
+  format(x, ..., n = NULL, width = NULL, n_extra = NULL)
+}
 
-  if (!is.null(src$con) && nrow(def) >= 0) {
+show_dm <- function(x) {
+  def <- dm_get_def(x)
+  if (nrow(def) == 0) {
+    cat_line("dm()")
+    return()
+  }
+
+  src <- dm_get_src(x)
+  if (!is.null(src)) {
+    cat_rule("Table source", col = "green")
+    db_info <- NULL
+
     # FIXME: change to pillar::tbl_sum() once it's there
     tbl_str <- tibble::tbl_sum(def$data[[1]])
     if ("Database" %in% names(tbl_str)) {
       db_info <- paste0("src:  ", tbl_str[["Database"]])
     }
-  }
-  if (is.null(db_info)) {
-    db_info <- strsplit(format(src), "\n")[[1]][[1]]
-  }
+    if (is.null(db_info)) {
+      db_info <- strsplit(format(src), "\n")[[1]][[1]]
+    }
 
-  cat_line(db_info)
+    cat_line(db_info)
+  }
 
   cat_rule("Metadata", col = "violet")
 
@@ -491,7 +502,23 @@ print.dm <- function(x, ...) {
     cat_rule("Filters", col = "orange")
     walk2(filters$table, filters$filter, ~ cat_line(paste0(.x, ": ", as_label(.y))))
   }
+}
 
+#' @export
+format.dm <- function(x, ...) { # for both dm and zoomed_dm
+  def <- dm_get_def(x)
+  glue("dm: {def_get_n_tables(def)} tables, {def_get_n_columns(def)} columns, {def_get_n_pks(def)} primary keys, {def_get_n_fks(def)} foreign keys")
+}
+
+#' @export
+format.zoomed_dm <- function(x, ..., n = NULL, width = NULL, n_extra = NULL) {
+  df <- get_zoomed_tbl(x)
+  # so far only 1 table can be zoomed on
+  zoomed_df <- new_zoomed_df(
+    df,
+    name_df = orig_name_zoomed(x)
+  )
+  cat_line(format(zoomed_df, ..., n = n, width = width, n_extra = n_extra))
   invisible(x)
 }
 
@@ -509,23 +536,6 @@ def_get_n_pks <- function(def) {
 
 def_get_n_fks <- function(def) {
   sum(map_int(def$fks, vctrs::vec_size))
-}
-
-#' @export
-print.zoomed_dm <- function(x, ..., n = NULL, width = NULL, n_extra = NULL) {
-  format(x, ..., n = NULL, width = NULL, n_extra = NULL)
-}
-
-#' @export
-format.zoomed_dm <- function(x, ..., n = NULL, width = NULL, n_extra = NULL) {
-  df <- get_zoomed_tbl(x)
-  # so far only 1 table can be zoomed on
-  zoomed_df <- new_zoomed_df(
-    df,
-    name_df = orig_name_zoomed(x)
-  )
-  cat_line(format(zoomed_df, ..., n = n, width = width, n_extra = n_extra))
-  invisible(x)
 }
 
 new_zoomed_df <- function(x, ...) {
@@ -561,15 +571,15 @@ format.zoomed_df <- function(x, ..., n = NULL, width = NULL, n_extra = NULL) {
 }
 
 #' @export
-`$.zoomed_dm` <- function(x, name) {
-  name <- ensym(name)
-  eval_tidy(quo(`$`(get_zoomed_tbl(x), !!name)))
+`$.dm` <- function(x, name) { # for both dm and zoomed_dm
+  table <- dm_tbl_name(x, {{ name }})
+  tbl(x, table)
 }
 
 #' @export
-`$.dm` <- function(x, name) {
-  table <- as_string(ensym(name))
-  tbl(x, table)
+`$.zoomed_dm` <- function(x, name) {
+  name <- ensym(name)
+  eval_tidy(quo(`$`(get_zoomed_tbl(x), !!name)))
 }
 
 #' @export
@@ -578,24 +588,19 @@ format.zoomed_df <- function(x, ..., n = NULL, width = NULL, n_extra = NULL) {
 }
 
 #' @export
-`[[.zoomed_dm` <- function(x, id) {
-  `[[`(get_zoomed_tbl(x), id)
-}
-
-#' @export
-`[[.dm` <- function(x, id) {
+`[[.dm` <- function(x, id) { # for both dm and zoomed_dm
   if (is.numeric(id)) id <- src_tbls(x)[id] else id <- as_string(id)
   tbl(x, id)
 }
 
 #' @export
-`[[<-.dm` <- function(x, name, value) {
-  abort_update_not_supported()
+`[[.zoomed_dm` <- function(x, id) {
+  `[[`(get_zoomed_tbl(x), id)
 }
 
 #' @export
-`[.zoomed_dm` <- function(x, id) {
-  `[`(get_zoomed_tbl(x), id)
+`[[<-.dm` <- function(x, name, value) {
+  abort_update_not_supported()
 }
 
 #' @export
@@ -605,6 +610,11 @@ format.zoomed_df <- function(x, ..., n = NULL, width = NULL, n_extra = NULL) {
   dm_select_tbl(x, !!!id)
 }
 
+#' @export
+`[.zoomed_dm` <- function(x, id) { # for both dm and zoomed_dm
+  `[`(get_zoomed_tbl(x), id)
+}
+
 
 #' @export
 `[<-.dm` <- function(x, name, value) {
@@ -612,13 +622,13 @@ format.zoomed_df <- function(x, ..., n = NULL, width = NULL, n_extra = NULL) {
 }
 
 #' @export
-names.zoomed_dm <- function(x) {
-  names(get_zoomed_tbl(x))
+names.dm <- function(x) { # for both dm and zoomed_dm
+  src_tbls(x)
 }
 
 #' @export
-names.dm <- function(x) {
-  src_tbls(x)
+names.zoomed_dm <- function(x) {
+  names(get_zoomed_tbl(x))
 }
 
 
@@ -628,13 +638,13 @@ names.dm <- function(x) {
 }
 
 #' @export
-length.zoomed_dm <- function(x) {
-  length(get_zoomed_tbl(x))
+length.dm <- function(x) { # for both dm and zoomed_dm
+  length(src_tbls(x))
 }
 
 #' @export
-length.dm <- function(x) {
-  length(src_tbls(x))
+length.zoomed_dm <- function(x) {
+  length(get_zoomed_tbl(x))
 }
 
 #' @export
@@ -643,7 +653,7 @@ length.dm <- function(x) {
 }
 
 #' @export
-str.dm <- function(object, ...) {
+str.dm <- function(object, ...) { # for both dm and zoomed_dm
   object <- dm_get_def(object) %>%
     select(table, pks, fks, filters)
   str(object)
@@ -666,10 +676,11 @@ str.zoomed_dm <- function(object, ...) {
 #' @rdname dplyr_db
 #' @export
 tbl.dm <- function(src, from, ...) {
+  check_not_zoomed(src)
+
   # The src argument here is a dm object
   dm <- src
-  check_not_zoomed(dm)
-  check_correct_input(dm, from, 1L)
+  from <- dm_tbl_name(dm, !!from)
 
   dm_get_tables_impl(dm)[[from]]
 }
@@ -677,7 +688,7 @@ tbl.dm <- function(src, from, ...) {
 #' @param x Either a `dm` or a `zoomed_dm`; the latter leads to an error for `src_tbls.dm()`
 #' @rdname dplyr_db
 #' @export
-compute.dm <- function(x, ...) {
+compute.dm <- function(x, ...) { # for both dm and zoomed_dm
   dm_apply_filters(x) %>%
     dm_get_def() %>%
     mutate(data = map(data, compute, ...)) %>%
@@ -696,10 +707,8 @@ compute.zoomed_dm <- function(x, ...) {
 #' @rdname dplyr_db
 #' @export
 src_tbls.dm <- function(x, ...) {
-  # The x argument here is a dm object
-  dm <- x
   check_not_zoomed(x)
-  src_tbls_impl(dm)
+  src_tbls_impl(x)
 }
 
 src_tbls_impl <- function(dm) {
@@ -716,14 +725,21 @@ src_tbls_impl <- function(dm) {
 #' @param repair,quiet Name repair options; cf. [`vctrs::vec_as_names`]
 #' @export
 copy_to.dm <- function(dest, df, name = deparse(substitute(df)), overwrite = FALSE, temporary = TRUE, repair = "unique", quiet = FALSE, ...) {
+  check_not_zoomed(dest)
+
   if (!(inherits(df, "data.frame") || inherits(df, "tbl_dbi"))) abort_only_data_frames_supported()
   if (overwrite) abort_no_overwrite()
   if (length(name) != 1) abort_one_name_for_copy_to(name)
   # src: if `df` on a different src:
   # if `df_list` is on DB and `dest` is local, collect `df_list`
   # if `df_list` is local and `dest` is on DB, copy `df_list` to respective DB
-  df <- copy_to(dm_get_src(dest), df, unique_db_table_name(name), temporary = temporary, ...)
-  # FIXME: should we allow `overwrite` argument?
+  dest_src <- dm_get_src(dest)
+  if (is.null(dest_src)) {
+    df <- as_tibble(collect(df))
+  } else {
+    # FIXME: should we allow `overwrite` argument?
+    df <- copy_to(dest_src, df, unique_db_table_name(name), temporary = temporary, ...)
+  }
   names_list <- repair_table_names(src_tbls(dest), name, repair, quiet)
   # rename old tables with potentially new names
   dest <- dm_rename_tbl(dest, !!!names_list$new_old_names)
@@ -731,14 +747,9 @@ copy_to.dm <- function(dest, df, name = deparse(substitute(df)), overwrite = FAL
   dm_add_tbl_impl(dest, list(df), names_list$new_names)
 }
 
-#' @export
-copy_to.zoomed_dm <- function(dest, df, name, overwrite, ...) {
-  check_not_zoomed(dest)
-}
-
 #' @rdname dplyr_db
 #' @export
-collect.dm <- function(x, ...) {
+collect.dm <- function(x, ...) { # for both dm and zoomed_dm
   x <- dm_apply_filters(x)
 
   def <- dm_get_def(x)
@@ -748,18 +759,20 @@ collect.dm <- function(x, ...) {
 
 #' @export
 collect.zoomed_dm <- function(x, ...) {
-  check_not_zoomed(x)
+  message("Detaching table from dm, use `collect(pull_tbl())` instead to silence this message.")
+
+  collect(pull_tbl(x))
 }
 
 
 # FIXME: what about 'dim.dm()'?
 #' @export
-dim.zoomed_dm <- function(x) {
+dim.zoomed_dm <- function(x) { # dm method provided by base
   dim(get_zoomed_tbl(x))
 }
 
 #' @export
-dimnames.zoomed_dm <- function(x) {
+dimnames.zoomed_dm <- function(x) { # dm method provided by base
   dimnames(get_zoomed_tbl(x))
 }
 
@@ -813,7 +826,7 @@ empty_dm <- function() {
 #'
 #' @return The requested table
 #'
-#' @examples
+#' @examplesIf rlang::is_installed("nycflights13")
 #' # For an unzoomed dm you need to specify the table to pull:
 #' dm_nycflights13() %>%
 #'   pull_tbl(airports)
@@ -828,7 +841,7 @@ pull_tbl <- function(dm, table) {
 }
 
 #' @export
-pull_tbl.dm <- function(dm, table) {
+pull_tbl.dm <- function(dm, table) { # for both dm and zoomed_dm
   # FIXME: shall we issue a special error in case someone tries sth. like: `pull_tbl(dm_for_filter, c(t4, t3))`?
   table_name <- as_string(enexpr(table))
   if (table_name == "") abort_no_table_provided()
@@ -840,7 +853,11 @@ pull_tbl.zoomed_dm <- function(dm, table) {
   table_name <- as_string(enexpr(table))
   tbl_zoomed <- dm_get_zoomed_tbl(dm)
   if (table_name == "") {
-    if (nrow(tbl_zoomed) == 1) tbl_zoomed$zoom[[1]] else abort_not_pulling_multiple_zoomed()
+    if (nrow(tbl_zoomed) == 1) {
+      tbl_zoomed$zoom[[1]]
+    } else {
+      abort_not_pulling_multiple_zoomed()
+    }
   } else if (!(table_name %in% tbl_zoomed$table)) {
     abort_table_not_zoomed(table_name, tbl_zoomed$table)
   } else {
@@ -851,7 +868,7 @@ pull_tbl.zoomed_dm <- function(dm, table) {
 }
 
 #' @export
-as.list.dm <- function(x, ...) {
+as.list.dm <- function(x, ...) { # for both dm and zoomed_dm
   dm_get_tables_impl(x)
 }