+++ tests/testthat/test-dm.R
@@ -1,4 +1,6 @@
 test_that("can access tables", {
+  skip_if_not_installed("nycflights13")
+
   expect_identical(tbl(dm_nycflights13(), "airlines"), nycflights13::airlines)
   expect_dm_error(
     tbl(dm_nycflights13(), "x"),
@@ -39,11 +41,7 @@ test_that("'copy_to.dm()' works", {
   skip_if_src_not("df", "mssql")
 
   # `tibble()` call necessary, #322
-  car_table <- copy_to(
-    my_test_src(),
-    tibble(mtcars),
-    name = unique_db_table_name("mtcars_1")
-  )
+  car_table <- test_src_frame(!!!mtcars)
 
   expect_equivalent_dm(
     copy_to(dm_for_filter(), mtcars, "car_table"),
@@ -61,21 +59,21 @@ test_that("'copy_to.dm()' works", {
 
 test_that("'copy_to.dm()' works (2)", {
   expect_dm_error(
-    copy_to(dm_for_filter(), mtcars, c("car_table", "another_table")),
+    copy_to(dm(), mtcars, c("car_table", "another_table")),
     "one_name_for_copy_to"
   )
 
   # rename old and new tables if `repair = unique`
   expect_name_repair_message(
     expect_equivalent_dm(
-      dm(mtcars) %>% copy_to(mtcars),
+      copy_to(dm(mtcars), mtcars),
       dm(mtcars...1 = mtcars, mtcars...2 = tibble(mtcars))
     )
   )
 
   expect_equivalent_dm(
     expect_silent(
-      dm(mtcars) %>% copy_to(mtcars, quiet = TRUE)
+      copy_to(dm(mtcars), mtcars, quiet = TRUE)
     ),
     dm(mtcars...1 = mtcars, mtcars...2 = tibble(mtcars))
   )
@@ -85,6 +83,8 @@ test_that("'copy_to.dm()' works (2)", {
     dm(mtcars) %>% copy_to(mtcars, repair = "check_unique"),
     "need_unique_names"
   )
+
+  skip_if_not_installed("dbplyr")
 
   # copying `tibble` from chosen src to sqlite() `dm`
   expect_equivalent_dm(
@@ -97,6 +97,31 @@ test_that("'copy_to.dm()' works (2)", {
     copy_to(dm_for_filter(), data_card_1_sqlite(), "test_table_1"),
     dm_add_tbl(dm_for_filter(), test_table_1 = data_card_1())
   )
+})
+
+test_that("'collect.dm()' collects tables on DB", {
+  def <-
+    dm_for_filter() %>%
+    dm_filter(tf_1, a > 3) %>%
+    collect() %>%
+    dm_get_def()
+
+  is_df <- map_lgl(def$data, is.data.frame)
+  expect_true(all(is_df))
+})
+
+test_that("'collect.zoomed_dm()' collects tables, with message", {
+  zoomed_dm_for_collect <-
+    dm_for_filter() %>%
+    dm_zoom_to(tf_1) %>%
+    mutate(c = a + 1)
+
+  expect_message(
+    out <- zoomed_dm_for_collect %>% collect(),
+    "pull_tbl"
+  )
+
+  expect_s3_class(out, "data.frame")
 })
 
 test_that("'compute.dm()' computes tables on DB", {
@@ -181,6 +206,8 @@ test_that("validator is silent", {
 })
 
 test_that("validator speaks up (sqlite())", {
+  skip_if_not_installed("dbplyr")
+
   expect_dm_error(
     new_dm3(dm_get_def(dm_for_filter()) %>%
       mutate(data = if_else(table == "tf_1", list(dm_for_filter_sqlite()$tf_1), data))) %>%
@@ -398,15 +425,17 @@ test_that("dm_get_filters() works", {
   )
 })
 
-test_that("output", {
+skip_if_not_installed("nycflights13")
+
+verify_output("out/output.txt", {
+  print(dm())
+
   nyc_flights_dm <- dm_nycflights13(cycle = TRUE)
-  verify_output("out/output.txt", {
-    nyc_flights_dm
+  nyc_flights_dm
 
-    nyc_flights_dm %>%
-      format()
+  nyc_flights_dm %>%
+    format()
 
-    nyc_flights_dm %>%
-      dm_filter(flights, origin == "EWR")
-  })
+  nyc_flights_dm %>%
+    dm_filter(flights, origin == "EWR")
 })